fix(Compiler): No longer emit "[WARNING]: Invalid @Injectable" incorrectly.
There was a convenient `_getCompileTypeMetadata` function that was used across
the view compiler/metadata collector, but it also checked for various states
and annotations (such as the class being creatable) that don't apply to all
cases.

For example, when using `useClass: Clazz`, you _need_ `Clazz` to be new-able
so it makes sense to emit a warning (an error would be better, but one thing
at a time). However, when you are just using an interface, and won't ever be
creating that class, the warning is wrong/never correct.

PiperOrigin-RevId: 186513446